pyproject.toml
[tool.poetry]
name = "DearMEP"
version = "0.0.1"
description = "DearMEP helps people contact their representatives."
authors = ["Tim Weber"]
license = "AGPL-3.0"

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.79"
prometheus-fastapi-instrumentator = "^5.9.1"
maxminddb = "^2.2.0"
PyYAML = "^6.0"

[tool.poetry.dev-dependencies]
uvicorn = {extras = ["standard"], version = "^0.18.2"}
flake8 = "^4.0.1"
flake8-quotes = "^3.3.1"
mypy = "^0.971"
pytest = "^7.1.2"
pytest-cov = "^3.0.0"
requests = "^2.28.1"
py-mmdb-encoder = "^1.0.4"
types-PyYAML = "^6.0.11"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
